Arregle Models en capa de proyecto

+++ Oracle.DataAccess/Models/Car.cs
@@ -12,7 +12,7 @@ namespace Oracle.DataAccess.Models
 
         public long Id { get; set; }
         public string Model { get; set; } = null!;
-        public int Year { get; set; }
+        public byte Year { get; set; }
         public string Transmission { get; set; } = null!;
         public string Color { get; set; } = null!;
